WIP: extracting ODMR
