+++ hardware/ni_x_series_scanner.py
@@ -26,14 +26,12 @@ import re
 import PyDAQmx as daq
 
 from core.module import Base, ConfigOption
-from interface.slow_counter_interface import SlowCounterInterface
 from interface.slow_counter_interface import SlowCounterConstraints
 from interface.slow_counter_interface import CountingMode
-from interface.odmr_counter_interface import ODMRCounterInterface
 from interface.confocal_scanner_interface import ConfocalScannerInterface
 
 
-class NationalInstrumentsXSeries(Base, ConfocalScannerInterface, ODMRCounterInterface):
+class NationalInstrumentsXSeriesScanner(Base, ConfocalScannerInterface):
     """ A National Instruments device that can count and control microvave generators.
 
     !!!!!! NI USB 63XX, NI PCIe 63XX and NI PXIe 63XX DEVICES ONLY !!!!!!
@@ -45,7 +43,7 @@ class NationalInstrumentsXSeries(Base, ConfocalScannerInterface, ODMRCounterInte
     Example config for copy-paste:
 
     nicard_6343:
-        module.Class: 'national_instruments_x_series.NationalInstrumentsXSeries'
+        module.Class: 'ni_x_series_scanner.NationalInstrumentsXSeriesScanner'
         photon_sources:
             - '/Dev1/PFI8'
         #    - '/Dev1/PFI9'
@@ -181,13 +179,11 @@ class NationalInstrumentsXSeries(Base, ConfocalScannerInterface, ODMRCounterInte
         retval = 0
         chanlist = [
             self._odmr_trigger_channel,
-            self._clock_channel,
             self._scanner_clock_channel,
             self._gate_in_channel
             ]
         chanlist.extend(self._scanner_ao_channels)
         chanlist.extend(self._photon_sources)
-        chanlist.extend(self._counter_channels)
         chanlist.extend(self._scanner_counter_channels)
 
         devicelist = []
@@ -1039,490 +1035,6 @@ class NationalInstrumentsXSeries(Base, ConfocalScannerInterface, ODMRCounterInte
         return -1 if a < 0 or b < 0 or c < 0 else 0
 
     # ================ End ConfocalScannerInterface Commands ===================
-
-    # ==================== ODMRCounterInterface Commands =======================
-    def set_up_odmr_clock(self, clock_frequency=None, clock_channel=None):
-        """ Configures the hardware clock of the NiDAQ card to give the timing.
-
-        @param float clock_frequency: if defined, this sets the frequency of
-                                      the clock
-        @param string clock_channel: if defined, this is the physical channel
-                                     of the clock
-
-        @return int: error code (0:OK, -1:error)
-        """
-
-        return self.set_up_scanner_clock(
-            clock_frequency=clock_frequency,
-            clock_channel=clock_channel,
-            idle=False)
-
-    def set_up_odmr(self, counter_channel=None, photon_source=None,
-                    clock_channel=None, odmr_trigger_channel=None):
-        """ Configures the actual counter with a given clock.
-
-        @param string counter_channel: if defined, this is the physical channel
-                                       of the counter
-        @param string photon_source: if defined, this is the physical channel
-                                     where the photons are to count from
-        @param string clock_channel: if defined, this specifies the clock for
-                                     the counter
-        @param string odmr_trigger_channel: if defined, this specifies the
-                                            trigger output for the microwave
-
-        @return int: error code (0:OK, -1:error)
-        """
-        if self._scanner_clock_daq_task is None and clock_channel is None:
-            self.log.error('No clock running, call set_up_clock before starting the counter.')
-            return -1
-        if len(self._scanner_counter_daq_tasks) > 0:
-            self.log.error('Another counter is already running, close this one first.')
-            return -1
-        if len(self._scanner_ai_channels) > 0 and self._scanner_analog_daq_task is not None:
-            self.log.error('Another analog is already running, close this one first.')
-            return -1
-
-        if clock_channel is not None:
-            my_clock_channel = clock_channel
-        else:
-            my_clock_channel = self._scanner_clock_channel
-
-        if counter_channel is not None:
-            my_counter_channel = counter_channel
-        else:
-            my_counter_channel = self._scanner_counter_channels[0]
-
-        if photon_source is not None:
-            my_photon_source = photon_source
-        else:
-            my_photon_source = self._photon_sources[0]
-
-        # this task will count photons with binning defined by the clock_channel
-        task = daq.TaskHandle()
-        if len(self._scanner_ai_channels) > 0:
-            atask = daq.TaskHandle()
-        try:
-            # create task for the counter
-            daq.DAQmxCreateTask('ODMRCounter', daq.byref(task))
-            if len(self._scanner_ai_channels) > 0:
-                daq.DAQmxCreateTask('ODMRAnalog', daq.byref(atask))
-
-            # set up semi period width measurement in photon ticks, i.e. the width
-            # of each pulse (high and low) generated by pulse_out_task is measured
-            # in photon ticks.
-            #   (this task creates a channel to measure the time between state
-            #    transitions of a digital signal and adds the channel to the task
-            #    you choose)
-            daq.DAQmxCreateCISemiPeriodChan(
-                # define to which task to# connect this function
-                task,
-                # use this counter channel
-                my_counter_channel,
-                # name to assign to it
-                'ODMR Counter',
-                # Expected minimum count value
-                0,
-                # Expected maximum count value
-                self._max_counts / self._scanner_clock_frequency,
-                # units of width measurement, here photon ticks
-                daq.DAQmx_Val_Ticks,
-                '')
-
-            # Analog task
-            if len(self._scanner_ai_channels) > 0:
-                daq.DAQmxCreateAIVoltageChan(
-                    atask,
-                    ', '.join(self._scanner_ai_channels),
-                    'ODMR Analog',
-                    daq.DAQmx_Val_RSE,
-                    -10,
-                    10,
-                    daq.DAQmx_Val_Volts,
-                    ''
-                )
-
-            # connect the pulses from the clock to the counter
-            daq.DAQmxSetCISemiPeriodTerm(
-                task,
-                my_counter_channel,
-                my_clock_channel + 'InternalOutput')
-
-            # define the source of ticks for the counter as self._photon_source
-            daq.DAQmxSetCICtrTimebaseSrc(
-                task,
-                my_counter_channel,
-                my_photon_source)
-
-            # start and stop pulse task to correctly initiate idle state high voltage.
-            daq.DAQmxStartTask(self._scanner_clock_daq_task)
-            # otherwise, it will be low until task starts, and MW will receive wrong pulses.
-            daq.DAQmxStopTask(self._scanner_clock_daq_task)
-
-            if self.lock_in_active:
-                ptask = daq.TaskHandle()
-                daq.DAQmxCreateTask('ODMRPulser', daq.byref(ptask))
-                daq.DAQmxCreateDOChan(
-                    ptask,
-                    '{0:s}, {1:s}'.format(self._odmr_trigger_line, self._odmr_switch_line),
-                    "ODMRPulserChannel",
-                    daq.DAQmx_Val_ChanForAllLines)
-
-                self._odmr_pulser_daq_task = ptask
-
-            # connect the clock to the trigger channel to give triggers for the
-            # microwave
-            daq.DAQmxConnectTerms(
-                self._scanner_clock_channel + 'InternalOutput',
-                self._odmr_trigger_channel,
-                daq.DAQmx_Val_DoNotInvertPolarity)
-            self._scanner_counter_daq_tasks.append(task)
-            if len(self._scanner_ai_channels) > 0:
-                self._scanner_analog_daq_task = atask
-        except:
-            self.log.exception('Error while setting up ODMR scan.')
-            return -1
-        return 0
-
-    def set_odmr_length(self, length=100):
-        """ Sets up the trigger sequence for the ODMR and the triggered microwave.
-
-        @param int length: length of microwave sweep in pixel
-
-        @return int: error code (0:OK, -1:error)
-        """
-        if len(self._scanner_counter_channels) > 0 and len(self._scanner_counter_daq_tasks) < 1:
-            self.log.error('No counter is running, cannot do ODMR without one.')
-            return -1
-
-        if len(self._scanner_ai_channels) > 0 and self._scanner_analog_daq_task is None:
-            self.log.error('No analog task is running, cannot do ODMR without one.')
-            return -1
-
-        self._odmr_length = length
-        try:
-            # set timing for odmr clock task to the number of pixel.
-            daq.DAQmxCfgImplicitTiming(
-                # define task
-                self._scanner_clock_daq_task,
-                # only a limited number of counts
-                daq.DAQmx_Val_FiniteSamps,
-                # count twice for each voltage +1 for starting this task.
-                # This first pulse will start the count task.
-                self._odmr_length + 1)
-
-            # set timing for odmr count task to the number of pixel.
-            daq.DAQmxCfgImplicitTiming(
-                # define task
-                self._scanner_counter_daq_tasks[0],
-                # only a limited number of counts
-                daq.DAQmx_Val_ContSamps,
-                # count twice for each voltage +1 for starting this task.
-                # This first pulse will start the count task.
-                2 * (self._odmr_length + 1))
-
-            # read samples from beginning of acquisition, do not overwrite
-            daq.DAQmxSetReadRelativeTo(
-                self._scanner_counter_daq_tasks[0],
-                daq.DAQmx_Val_CurrReadPos)
-
-            # do not read first sample
-            daq.DAQmxSetReadOffset(
-                self._scanner_counter_daq_tasks[0],
-                0)
-
-            # unread data in buffer will be overwritten
-            daq.DAQmxSetReadOverWrite(
-                self._scanner_counter_daq_tasks[0],
-                daq.DAQmx_Val_DoNotOverwriteUnreadSamps)
-
-            # Analog
-            if len(self._scanner_ai_channels) > 0:
-                # Analog in channel timebase
-                daq.DAQmxCfgSampClkTiming(
-                    self._scanner_analog_daq_task,
-                    self._scanner_clock_channel + 'InternalOutput',
-                    self._scanner_clock_frequency,
-                    daq.DAQmx_Val_Rising,
-                    daq.DAQmx_Val_ContSamps,
-                    self._odmr_length + 1
-                )
-
-            if self._odmr_pulser_daq_task:
-                # pulser channel timebase
-                daq.DAQmxCfgSampClkTiming(
-                    self._odmr_pulser_daq_task,
-                    self._scanner_clock_channel + 'InternalOutput',
-                    self._scanner_clock_frequency,
-                    daq.DAQmx_Val_Rising,
-                    daq.DAQmx_Val_ContSamps,
-                    self._odmr_length + 1
-                )
-        except:
-            self.log.exception('Error while setting up ODMR counter.')
-            return -1
-        return 0
-
-    @property
-    def oversampling(self):
-        return self._oversampling
-
-    @oversampling.setter
-    def oversampling(self, val):
-        if not isinstance(val, (int, float)):
-            self.log.error('oversampling has to be int of float.')
-        else:
-            self._oversampling = int(val)
-
-    @property
-    def lock_in_active(self):
-        return self._lock_in_active
-
-    @lock_in_active.setter
-    def lock_in_active(self, val):
-        if not isinstance(val, bool):
-            self.log.error('lock_in_active has to be boolean.')
-        else:
-            self._lock_in_active = val
-            if self._lock_in_active:
-                self.log.warn('You just switched the ODMR counter to Lock-In-mode. \n'
-                              'Please make sure you connected all triggers correctly:\n'
-                              '  {0:s} is the microwave trigger channel\n'
-                              '  {1:s} is the switching channel for the lock in\n'
-                              ''.format(self._odmr_trigger_line, self._odmr_switch_line))
-
-    def count_odmr(self, length=100):
-        """ Sweeps the microwave and returns the counts on that sweep.
-
-        @param int length: length of microwave sweep in pixel
-
-        @return float[]: the photon counts per second
-        """
-        if len(self._scanner_counter_daq_tasks) < 1:
-            self.log.error(
-                'No counter is running, cannot scan an ODMR line without one.')
-            return True, np.array([-1.])
-
-        if len(self._scanner_ai_channels) > 0 and self._scanner_analog_daq_task is None:
-            self.log.error('No analog task is running, cannot do ODMR without one.')
-            return True, np.array([-1.])
-
-        # check if length setup is correct, if not, adjust.
-        if self._odmr_pulser_daq_task:
-            odmr_length_to_set = length * self.oversampling * 2
-        else:
-            odmr_length_to_set = length
-
-        if self.set_odmr_length(odmr_length_to_set) < 0:
-            self.log.error('An error arose while setting the odmr lenth to {}.'.format(odmr_length_to_set))
-            return True, np.array([-1.])
-
-        try:
-            # start the scanner counting task that acquires counts synchronously
-            daq.DAQmxStartTask(self._scanner_counter_daq_tasks[0])
-            if len(self._scanner_ai_channels) > 0:
-                daq.DAQmxStartTask(self._scanner_analog_daq_task)
-        except:
-            self.log.exception('Cannot start ODMR counter.')
-            return True, np.array([-1.])
-
-        if self._odmr_pulser_daq_task:
-            try:
-
-                # The pulse pattern is an alternating 0 and 1 on the switching channel (line0),
-                # while the first half of the whole microwave pulse is 1 and the other half is 0.
-                # This way the beginning of the microwave has a rising edge.
-                pulse_pattern = np.zeros(self.oversampling * 2, dtype=np.uint32)
-                pulse_pattern[:self.oversampling] += 1
-                pulse_pattern[::2] += 2
-
-                daq.DAQmxWriteDigitalU32(self._odmr_pulser_daq_task,
-                                         len(pulse_pattern),
-                                         0,
-                                         self._RWTimeout * self._odmr_length,
-                                         daq.DAQmx_Val_GroupByChannel,
-                                         pulse_pattern,
-                                         None,
-                                         None)
-
-                daq.DAQmxStartTask(self._odmr_pulser_daq_task)
-            except:
-                self.log.exception('Cannot start ODMR pulser.')
-                return True, np.array([-1.])
-
-        try:
-            daq.DAQmxStartTask(self._scanner_clock_daq_task)
-
-            # wait for the scanner clock to finish
-            daq.DAQmxWaitUntilTaskDone(
-                # define task
-                self._scanner_clock_daq_task,
-                # maximal timeout for the counter times the positions
-                self._RWTimeout * 2 * self._odmr_length)
-
-            # count data will be written here
-            odmr_data = np.full(
-                (2 * self._odmr_length + 1, ),
-                222,
-                dtype=np.uint32)
-
-            #number of samples which were read will be stored here
-            n_read_samples = daq.int32()
-
-            # actually read the counted photons
-            daq.DAQmxReadCounterU32(
-                # read from this task
-                self._scanner_counter_daq_tasks[0],
-                # Read number of double the# number of samples
-                2 * self._odmr_length + 1,
-                # Maximal timeout for the read # process
-                self._RWTimeout,
-                # write into this array
-                odmr_data,
-                # length of array to write into
-                2 * self._odmr_length + 1,
-                # number of samples which were actually read
-                daq.byref(n_read_samples),
-                # Reserved for future use. Pass NULL (here None) to this parameter.
-                None)
-
-            # Analog
-            if len(self._scanner_ai_channels) > 0:
-                odmr_analog_data = np.full(
-                    (len(self._scanner_ai_channels), self._odmr_length + 1),
-                    222,
-                    dtype=np.float64)
-
-                analog_read_samples = daq.int32()
-
-                daq.DAQmxReadAnalogF64(
-                    self._scanner_analog_daq_task,
-                    self._odmr_length + 1,
-                    self._RWTimeout,
-                    daq.DAQmx_Val_GroupByChannel,
-                    odmr_analog_data,
-                    len(self._scanner_ai_channels) * (self._odmr_length + 1),
-                    daq.byref(analog_read_samples),
-                    None
-                )
-
-            # stop the counter task
-            daq.DAQmxStopTask(self._scanner_clock_daq_task)
-            daq.DAQmxStopTask(self._scanner_counter_daq_tasks[0])
-            if len(self._scanner_ai_channels) > 0:
-                daq.DAQmxStopTask(self._scanner_analog_daq_task)
-            if self._odmr_pulser_daq_task:
-                daq.DAQmxStopTask(self._odmr_pulser_daq_task)
-
-            # prepare array to return data
-            all_data = np.full((len(self.get_odmr_channels()), length),
-                               222,
-                               dtype=np.float64)
-
-            # create a new array for the final data (this time of the length
-            # number of samples)
-            real_data = np.zeros((self._odmr_length, ), dtype=np.uint32)
-
-            # add upp adjoint pixels to also get the counts from the low time of
-            # the clock:
-
-            real_data += odmr_data[1:-1:2]
-            real_data += odmr_data[:-1:2]
-
-            if self._odmr_pulser_daq_task:
-                differential_data = np.zeros((self.oversampling * length, ), dtype=np.float64)
-
-                differential_data += real_data[1::2]
-                differential_data -= real_data[::2]
-                differential_data = np.divide(differential_data, real_data[::2],
-                                              np.zeros_like(differential_data),
-                                              where=real_data[::2] != 0)
-
-                all_data[0] = np.median(np.reshape(differential_data,
-                                                   (-1, self.oversampling)),
-                                        axis=1
-                                        )
-
-                if len(self._scanner_ai_channels) > 0:
-                    for i, analog_data in enumerate(odmr_analog_data):
-                        differential_data = np.zeros((self.oversampling * length, ), dtype=np.float64)
-
-                        differential_data += analog_data[1:-1:2]
-                        differential_data -= analog_data[:-1:2]
-                        differential_data = np.divide(differential_data, analog_data[:-1:2],
-                                                      np.zeros_like(differential_data),
-                                                      where=analog_data[:-1:2] != 0)
-
-                        all_data[i+1] = np.median(np.reshape(differential_data,
-                                                             (-1, self.oversampling)),
-                                                  axis=1
-                                                  )
-
-            else:
-                all_data[0] = np.array(real_data * self._scanner_clock_frequency, np.float64)
-                if len(self._scanner_ai_channels) > 0:
-                    all_data[1:] = odmr_analog_data[:, :-1]
-
-            return False, all_data
-        except:
-            self.log.exception('Error while counting for ODMR.')
-            return True, np.full((len(self.get_odmr_channels()), 1), [-1.])
-
-    def close_odmr(self):
-        """ Closes the odmr and cleans up afterwards.
-
-        @return int: error code (0:OK, -1:error)
-        """
-        retval = 0
-        try:
-            # disconnect the trigger channel
-            daq.DAQmxDisconnectTerms(
-                self._scanner_clock_channel + 'InternalOutput',
-                self._odmr_trigger_channel)
-
-        except:
-            self.log.exception('Error while disconnecting ODMR clock channel.')
-            retval = -1
-
-        if len(self._scanner_ai_channels) > 0:
-            try:
-                # stop the counter task
-                daq.DAQmxStopTask(self._scanner_analog_daq_task)
-                # after stopping delete all the configuration of the counter
-                daq.DAQmxClearTask(self._scanner_analog_daq_task)
-                # set the task handle to None as a safety
-                self._scanner_analog_daq_task = None
-            except:
-                self.log.exception('Could not close analog.')
-                retval = -1
-
-        if self._odmr_pulser_daq_task:
-            try:
-                # stop the pulser task
-                daq.DAQmxStopTask(self._odmr_pulser_daq_task)
-                # after stopping delete all the configuration of the pulser
-                daq.DAQmxClearTask(self._odmr_pulser_daq_task)
-                # set the task handle to None as a safety
-                self._odmr_pulser_daq_task = None
-            except:
-                self.log.exception('Could not close pulser.')
-                retval = -1
-
-        retval = -1 if self.close_counter(scanner=True) < 0 or retval < 0 else 0
-        return retval
-
-    def get_odmr_channels(self):
-        ch = [self._scanner_counter_channels[0]]
-        ch.extend(self._scanner_ai_channels)
-        return ch
-
-    def close_odmr_clock(self):
-        """ Closes the odmr and cleans up afterwards.
-
-        @return int: error code (0:OK, -1:error)
-        """
-        return self.close_scanner_clock(scanner=True)
-
-    # ================== End ODMRCounterInterface Commands ====================
 
     def get_status(self):
         """ Receives the current status of the Fast Counter and outputs it as
